+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Unreleased
+    * Fix `PrintStr.h` for STM32duino by removing `override` in `flush()`.
 * 1.1.2 (2020-11-28)
     * Minor header include fix for compatility with UnixHostDuino.
     * Fix various compilation warnings and errors on AVR compiler.