+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Unreleased
+* 1.2 (2021-01-04)
     * Fix `PrintStr.h` for STM32duino by removing `override` in `flush()`.
     * Add `hashDjb2()` hash function for normal and flash strings.
 * 1.1.2 (2020-11-28)